requirements.txt
envyaml
fastapi~=0.63
uvicorn~=0.13
uvloop~=0.15
pydantic~=1.8.1
sentry-sdk~=1.0
sqlalchemy~=1.4
alembic~=1.6
jinja2~=2.11
aiofiles~=0.6.0
python-multipart~=0.0.5
python-jose[cryptography]~=3.2
passlib[bcrypt]~=1.7
